CASi: allowing double speed factor
git-svn-id: https://dev.imis.uni-luebeck.de/svn/mate/trunk/CASi@1409 042ff3d9-7651-4cbe-9b35-41cc136bb496

+++ CASi/src/de/uniluebeck/imis/casi/ParameterParser.java
@@ -38,7 +38,7 @@ public class ParameterParser {
 	/** Should the gui be disabled? */
 	private boolean disableGui;
 	/** A factor for the simulation speed */
-	private int speedFactor;
+	private double speedFactor;
 	/** The configuration file for the network if one was connected */
 	private String networkConfigFile;
 	/** The format for log files */
@@ -59,7 +59,7 @@ public class ParameterParser {
 			if (args[i].matches("--network-config") && args.length > i + 1) {
 				networkConfigFile = args[++i];
 			} else if (args[i].matches("--speed") && args.length > i + 1) {
-				speedFactor = Integer.parseInt(args[++i]);
+				speedFactor = Double.parseDouble(args[++i]);
 			} else if (args[i].matches("--help")) {
 				helpRequest = true;
 			} else {
@@ -170,7 +170,7 @@ public class ParameterParser {
 		return logFormat;
 	}
 	
-	public int getSpeedFactor() {
+	public double getSpeedFactor() {
 		return speedFactor;
 	}
